+++ packages/manager/apps/dedicated/client/app/license/order/options/plesk/controller.js
@@ -1,0 +1,12 @@
+export default class PleskLicenseAdditionalOptionsController {
+  $onInit() {
+    this.powerpack = false;
+  }
+
+  onPowerpackOptionChange(value) {
+    this.options.PLESK.powerpack = value
+      ? { value: this.version.more.powerPackPlanCode }
+      : null;
+    this.onOptionsChange();
+  }
+}