+++ packages/manager/core/ovh-product-icons/icons/telecomIcons.ts
@@ -1,0 +1,8 @@
+import { TelecomEthernet, Telephony } from '../assets-icon-svg/icons';
+
+const telecomIcons = {
+  TelecomEthernet,
+  Telephony,
+};
+
+export default telecomIcons;