+++ packages/manager/core/ovh-themes/ui-kit/foundations/radius.ts
@@ -1,5 +1,0 @@
-export default {
-  small: '2px',
-  medium: '6px',
-  large: '8px',
-};