+++ packages/manager/core/ovh-themes/ui-kit/foundations/shadows.ts
@@ -1,8 +1,0 @@
-export default {
-  // This is none because it's not the real outline property
-  // it's just a shadow on focus-visible.
-  // Outline styles are overriden in global styles.
-  outline: 'none',
-  primary: '0 0 6px 0 rgba(0, 14, 156, 0.2)',
-  secondary: '0 4px 6px 0 rgba(0, 14, 156, 0.2)',
-};