+++ packages/manager/core/ovh-themes/ui-kit/ovh-components/icons/system/ArrowDownIcon.ts
@@ -1,7 +1,0 @@
-import { createIcon } from '@chakra-ui/react';
-
-export const ArrowDownIcon = createIcon({
-  displayName: 'ArrowDownIcon',
-  viewBox: '0 0 24 24',
-  d: 'M12 2a1 1 0 0 1 1 1l-.001 15.586 4.294-4.293a1 1 0 0 1 1.414 1.414L12 22.414l-.7-.7a1.006 1.006 0 0 1-.015-.014l-5.992-5.993a1 1 0 0 1 1.414-1.414L11 18.584V3a1 1 0 0 1 1-1z',
-});