+++ packages/manager/core/ovh-themes/ui-kit/ovh-components/icons/system/ChevronRightIcon.ts
@@ -1,7 +1,0 @@
-import { createIcon } from '@chakra-ui/react';
-
-export const ChevronRightIcon = createIcon({
-  displayName: 'ChevronRightIcon',
-  viewBox: '0 0 24 24',
-  d: 'M9.086 6.914A1 1 0 0 1 10.5 5.5l6.707 6.707-6.707 6.707A1 1 0 0 1 9.086 17.5l5.293-5.293-5.293-5.293z',
-});