+++ packages/manager/core/ovh-themes/ui-kit/ovh-components/icons/system/EllipsisIcon.ts
@@ -1,7 +1,0 @@
-import { createIcon } from '@chakra-ui/react';
-
-export const EllipsisIcon = createIcon({
-  displayName: 'EllipsisIcon',
-  viewBox: '0 0 24 24',
-  d: 'M12 10a2 2 0 1 1 0 4 2 2 0 0 1 0-4zm-6 0a2 2 0 1 1 0 4 2 2 0 0 1 0-4zm12 0a2 2 0 1 1 0 4 2 2 0 0 1 0-4z',
-});