+++ packages/manager/core/ovh-themes/ui-kit/ovh-components/icons/system/SortIcon.ts
@@ -1,7 +1,0 @@
-import { createIcon } from '@chakra-ui/react';
-
-export const SortIcon = createIcon({
-  displayName: 'SortIcon',
-  viewBox: '0 0 24 24',
-  d: 'M16.183 16a1.166 1.166 0 0 1 .825 1.992l-4.183 4.183a1.166 1.166 0 0 1-1.65 0l-4.183-4.183A1.166 1.166 0 0 1 7.817 16zM11.175 5.825a1.166 1.166 0 0 1 1.65 0l4.183 4.183A1.166 1.166 0 0 1 16.183 12H7.817a1.166 1.166 0 0 1-.825-1.992z',
-});