+++ packages/manager/modules/pci/src/components/index.js
@@ -4,6 +4,7 @@ import cucConsumption from './consumption';
 import cucCurrency from './currency';
 import project from './project';
 import projectQuotaExceedErrorModal from './project-quota-exceed-error-modal';
+import pciAiDataSyncModal from './pci-ai-data-sync';
 import runabove from './runabove';
 import writeRightRequired from './writeRightRequired';
 import ovhManagerPciOrderCommand from './pci-order-command';
@@ -18,6 +19,7 @@ angular.module(moduleName, [
   runabove,
   writeRightRequired,
   ovhManagerPciOrderCommand,
+  pciAiDataSyncModal,
 ]);
 
 export default moduleName;