+++ packages/manager/modules/pci/src/projects/project/storages/databases/add/add.controller.js
@@ -189,7 +189,10 @@ export default class {
   }
 
   updateRegion(region) {
-    this.model.region = region;
+    if (this.model.region.name !== region.name) {
+      this.model.region = region;
+      this.model.privateNetwork = null;
+    }
     this.model.flavor = this.getSyncFlavor(region);
     this.updateFlavor(this.model.flavor);
   }
@@ -205,7 +208,10 @@ export default class {
       this.model.usePrivateNetwork = false;
       this.model.subnet = null;
       this.model.privateNetwork = null;
-    } else if (!this.model.usePrivateNetwork) {
+    } else if (
+      !this.model.usePrivateNetwork ||
+      this.model.privateNetwork === null
+    ) {
       this.model.privateNetwork = this.defaultPrivateNetwork;
       this.model.subnet = null;