+++ packages/manager/core/generator/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-generator@0.4.0...@ovh-ux/manager-generator@0.5.0) (2024-01-08)
+
+
+### Features
+
+* **generator:** add pagination v2 cursor and fix pagination v6 ([#10379](https://github.com/ovh/manager/issues/10379)) ([4fd7800](https://github.com/ovh/manager/commit/4fd7800c51e0fc224520e9c90126879c4c550247))
+
+
+
+
+
 # [0.4.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-generator@0.3.0...@ovh-ux/manager-generator@0.4.0) (2023-12-19)