+++ packages/manager/modules/emailpro/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.10.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-emailpro@3.9.2...@ovh-ux/manager-emailpro@3.10.0) (2024-01-03)
+
+
+### Features
+
+* **web:** manage DKIM key in the manager ([#10218](https://github.com/ovh/manager/issues/10218)) ([0e469cd](https://github.com/ovh/manager/commit/0e469cd120045c7800d84823f6915697f052f961))
+
+
+
+
+
 ## [3.9.2](https://github.com/ovh/manager/compare/@ovh-ux/manager-emailpro@3.9.1...@ovh-ux/manager-emailpro@3.9.2) (2024-01-02)